+++ holiday/src/main/java/org/kernely/holiday/dto/HolidayDetailCreationRequestDTO.java
@@ -28,7 +28,7 @@ public class HolidayDetailCreationRequestDTO {
 	/**
 	 * The id of holiday type
 	 */
-	public int typeInstanceId;
+	public long typeInstanceId;
 
 	/**
 	 * Default constructor
